+++ Cargo.toml
@@ -5,7 +5,6 @@ members = [
     "cumulus/client/block-builder",
     "cumulus/client/cli",
     "cumulus/client/cirrus-executor",
-    "cumulus/client/consensus/common",
     "cumulus/client/consensus/relay-chain",
     "cumulus/client/executor-gossip",
     "cumulus/pallets/executive",
