--- cumulus/client/consensus/common/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "cumulus-client-consensus-common"
-description = "Cumulus specific common consensus implementations"
-version = "0.1.0"
-authors = ["Parity Technologies <admin@parity.io>"]
-edition = "2021"
-
-[dependencies]
-async-trait = "0.1.53"
-sc-consensus = { git = "https://github.com/paritytech/substrate", rev = "c4f3d028621edb293d2c423516221aa396f76a2d" }
-sp-consensus = { git = "https://github.com/paritytech/substrate", rev = "c4f3d028621edb293d2c423516221aa396f76a2d" }
-sp-runtime = { git = "https://github.com/paritytech/substrate", rev = "c4f3d028621edb293d2c423516221aa396f76a2d" }
